requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==1.10.13
requests==2.31.0
twilio==8.10.0

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# AI/ML
google-generativeai==0.3.2

# Frontend Dependencies
streamlit==1.28.1
streamlit-lottie==0.0.5

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
